set up account view and login correctly
