+++ app/auth/signout/route.ts
@@ -1,0 +1,21 @@
+import { createClient } from "@/utils/supabase/server";
+import { revalidatePath } from "next/cache";
+import { type NextRequest, NextResponse } from "next/server";
+
+export async function POST(req: NextRequest) {
+ const supabase = await createClient();
+
+ // Check if a user's logged in
+ const {
+  data: { user },
+ } = await supabase.auth.getUser();
+
+ if (user) {
+  await supabase.auth.signOut();
+ }
+
+ revalidatePath("/", "layout");
+ return NextResponse.redirect(new URL("/login", req.url), {
+  status: 302,
+ });
+}
