+++ app/dashboard/layout.tsx
@@ -3,101 +3,18 @@
 import type React from "react";
 
 import { useState } from "react";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Badge } from "@/components/ui/badge";
-import {
- DropdownMenu,
- DropdownMenuContent,
- DropdownMenuItem,
- DropdownMenuLabel,
- DropdownMenuSeparator,
- DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
-import { Sheet, SheetContent, SheetTrigger } from "@/components/ui/sheet";
-import {
- Home,
- Users,
- Calendar,
- ChefHat,
- MessageSquare,
- BarChart3,
- Settings,
- HelpCircle,
- Bell,
- Search,
- Menu,
- LogOut,
- User,
-} from "lucide-react";
-import Link from "next/link";
-import { usePathname } from "next/navigation";
-// import { useSearchParams } from "next/navigation"
+import { Sheet, SheetContent } from "@/components/ui/sheet";
 import { Suspense } from "react";
-
-const navigation = [
- { name: "Dashboard", href: "/dashboard", icon: Home },
- { name: "Clients", href: "/dashboard/clients", icon: Users },
- { name: "Appointments", href: "/dashboard/appointments", icon: Calendar },
- { name: "Meal Plans", href: "/dashboard/meal-plans", icon: ChefHat },
- { name: "Messages", href: "/dashboard/messages", icon: MessageSquare },
- { name: "Analytics", href: "/dashboard/analytics", icon: BarChart3 },
- { name: "Settings", href: "/dashboard/settings", icon: Settings },
-];
+import Sidebar from "@/components/molecules/Sidebar";
+import DashboardHeader from "@/components/organisms/dashboard/DashboardHeader";
 
 export default function DashboardLayout({
  children,
 }: {
  children: React.ReactNode;
 }) {
- const pathname = usePathname();
  // const searchParams = useSearchParams()
  const [sidebarOpen, setSidebarOpen] = useState(false);
-
- const Sidebar = ({ mobile = false }) => (
-  <div className={`flex flex-col h-full ${mobile ? "w-full" : "w-64"}`}>
-   {/* Logo */}
-   <div className="flex items-center space-x-2 p-6 border-b">
-    <div className="w-8 h-8 bg-[#1B5E20] rounded-lg flex items-center justify-center">
-     <ChefHat className="w-5 h-5 text-white" />
-    </div>
-    <span className="text-xl font-bold text-[#1B5E20]">AfyaNutrix</span>
-   </div>
-
-   {/* Navigation */}
-   <nav className="flex-1 p-4 space-y-2">
-    {navigation.map((item) => {
-     const isActive = pathname === item.href;
-     return (
-      <Link
-       key={item.name}
-       href={item.href}
-       className={`flex items-center space-x-3 px-3 py-2 rounded-lg text-sm font-medium transition-colors ${
-        isActive
-         ? "bg-[#1B5E20] text-white"
-         : "text-gray-600 hover:bg-gray-100 hover:text-gray-900"
-       }`}
-       onClick={() => mobile && setSidebarOpen(false)}
-      >
-       <item.icon className="w-5 h-5" />
-       <span>{item.name}</span>
-      </Link>
-     );
-    })}
-   </nav>
-
-   {/* Help Section */}
-   <div className="p-4 border-t">
-    <Link
-     href="/help"
-     className="flex items-center space-x-3 px-3 py-2 rounded-lg text-sm font-medium text-gray-600 hover:bg-gray-100 hover:text-gray-900 transition-colors"
-    >
-     <HelpCircle className="w-5 h-5" />
-     <span>Help & Support</span>
-    </Link>
-   </div>
-  </div>
- );
 
  return (
   <Suspense fallback={<div>Loading...</div>}>
@@ -105,7 +22,7 @@ export default function DashboardLayout({
     {/* Desktop Sidebar */}
     <div className="hidden lg:fixed lg:inset-y-0 lg:flex lg:w-64 lg:flex-col">
      <div className="flex flex-col flex-grow bg-white border-r border-gray-200 overflow-y-auto">
-      <Sidebar />
+      <Sidebar sidebarOpen={sidebarOpen} setSidebarOpen={setSidebarOpen} />
      </div>
     </div>
 
@@ -113,86 +30,21 @@ export default function DashboardLayout({
     <Sheet open={sidebarOpen} onOpenChange={setSidebarOpen}>
      <SheetContent side="left" className="p-0 w-64">
       <div className="bg-white h-full">
-       <Sidebar mobile />
+       <Sidebar
+        mobile
+        sidebarOpen={sidebarOpen}
+        setSidebarOpen={setSidebarOpen}
+       />
       </div>
      </SheetContent>
     </Sheet>
 
     {/* Main Content */}
     <div className="lg:pl-64 flex flex-col flex-1">
-     {/* Top Header */}
-     <header className="bg-white border-b border-gray-200 px-4 py-4 sm:px-6 lg:px-8">
-      <div className="flex items-center justify-between">
-       <div className="flex items-center space-x-4">
-        <Sheet>
-         <SheetTrigger asChild>
-          <Button variant="ghost" size="sm" className="lg:hidden">
-           <Menu className="w-5 h-5" />
-          </Button>
-         </SheetTrigger>
-         <SheetContent side="left" className="p-0 w-64">
-          <div className="bg-white h-full">
-           <Sidebar mobile />
-          </div>
-         </SheetContent>
-        </Sheet>
-
-        {/* Search */}
-        <div className="relative hidden sm:block">
-         <Search className="absolute left-3 top-1/2 transform -translate-y-1/2 text-gray-400 w-4 h-4" />
-         <Input
-          placeholder="Search clients, appointments..."
-          className="pl-10 w-64"
-         />
-        </div>
-       </div>
-
-       <div className="flex items-center space-x-4">
-        {/* Notifications */}
-        <Button variant="ghost" size="sm" className="relative">
-         <Bell className="w-5 h-5" />
-         <Badge className="absolute -top-1 -right-1 w-5 h-5 p-0 flex items-center justify-center text-xs bg-red-500">
-          3
-         </Badge>
-        </Button>
-
-        {/* User Menu */}
-        <DropdownMenu>
-         <DropdownMenuTrigger asChild>
-          <Button variant="ghost" className="flex items-center space-x-2">
-           <div className="w-8 h-8 bg-[#1B5E20] rounded-full flex items-center justify-center text-white text-sm font-medium">
-            A
-           </div>
-           <span className="hidden sm:block text-sm font-medium">
-            Amina Wanjiku
-           </span>
-          </Button>
-         </DropdownMenuTrigger>
-         <DropdownMenuContent align="end" className="w-56">
-          <DropdownMenuLabel>My Account</DropdownMenuLabel>
-          <DropdownMenuSeparator />
-          <DropdownMenuItem>
-           <User className="mr-2 h-4 w-4" />
-           <span>Profile</span>
-          </DropdownMenuItem>
-          <DropdownMenuItem>
-           <Settings className="mr-2 h-4 w-4" />
-           <span>Settings</span>
-          </DropdownMenuItem>
-          <DropdownMenuItem>
-           <HelpCircle className="mr-2 h-4 w-4" />
-           <span>Help</span>
-          </DropdownMenuItem>
-          <DropdownMenuSeparator />
-          <DropdownMenuItem className="text-red-600">
-           <LogOut className="mr-2 h-4 w-4" />
-           <span>Log out</span>
-          </DropdownMenuItem>
-         </DropdownMenuContent>
-        </DropdownMenu>
-       </div>
-      </div>
-     </header>
+     <DashboardHeader
+      sidebarOpen={sidebarOpen}
+      setSidebarOpen={setSidebarOpen}
+     />
 
      {/* Page Content */}
      <main className="flex-1 p-4 sm:p-6 lg:p-8">{children}</main>
